Edit module system
