+++ js/data.js
@@ -1,38 +1,37 @@
-const ITEMS_COUNT = 25;
-
-const MESSAGES = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
-];
-
-const NAMES = [
-  'Артем',
-  'Саша',
-  'Паша',
-  'Дмитрий',
-  'Сергей',
-  'Алина',
-  'Светлана',
-  'Надежда',
-  'Ольга',
-  'Харима'
-];
-
-const DESCRIPTIONS = [
-  'Восход солнца в горах - невероятное зрелище, которое заставляет сердце биться быстрее.',
-  'Моя собака на прогулке - истинный образец верности и беззаветной любви.',
-  'Путешествие по Европе: волшебные улочки Праги захватывают дух своей красотой и атмосферой.',
-  'Семейный ужин в ресторане с видом на закат - момент счастья и спокойствия.',
-  'Мой новый тату: символ моей силы и решимости, который всегда будет напоминать мне о ценности жизни.',
-  'Спортивный тренировочный процесс - борьба с собой и постоянное стремление к лучшему.',
-  'Романтический ужин на пляже под звездами - мгновение, когда время останавливается.',
-  'Мои дети на празднике в садике - радость и бесконечная энергия, которые наполняют мою жизнь смыслом.',
-  'Фотография города из окна самолета - новые горизонты и возможности, которые ждут меня впереди.',
-  'Мой новый автомобиль - символ достижения моих целей и трудолюбия, который стал реальностью благодаря усердной работе.'
-];
-
-export {ITEMS_COUNT, MESSAGES, NAMES, DESCRIPTIONS};
+const getDataForPosts = () => {
+  const ITEMS_COUNT = 25;
+  const MESSAGES = [
+    'Всё отлично!',
+    'В целом всё неплохо. Но не всё.',
+    'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+    'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+    'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+    'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+  ];
+  const NAMES = [
+    'Артем',
+    'Саша',
+    'Паша',
+    'Дмитрий',
+    'Сергей',
+    'Алина',
+    'Светлана',
+    'Надежда',
+    'Ольга',
+    'Харима'
+  ];
+  const DESCRIPTIONS = [
+    'Восход солнца в горах - невероятное зрелище, которое заставляет сердце биться быстрее.',
+    'Моя собака на прогулке - истинный образец верности и беззаветной любви.',
+    'Путешествие по Европе: волшебные улочки Праги захватывают дух своей красотой и атмосферой.',
+    'Семейный ужин в ресторане с видом на закат - момент счастья и спокойствия.',
+    'Мой новый тату: символ моей силы и решимости, который всегда будет напоминать мне о ценности жизни.',
+    'Спортивный тренировочный процесс - борьба с собой и постоянное стремление к лучшему.',
+    'Романтический ужин на пляже под звездами - мгновение, когда время останавливается.',
+    'Мои дети на празднике в садике - радость и бесконечная энергия, которые наполняют мою жизнь смыслом.',
+    'Фотография города из окна самолета - новые горизонты и возможности, которые ждут меня впереди.',
+    'Мой новый автомобиль - символ достижения моих целей и трудолюбия, который стал реальностью благодаря усердной работе.'
+  ];
+  return {ITEMS_COUNT, MESSAGES, NAMES, DESCRIPTIONS};
+};
+export {getDataForPosts};
